+++ CHANGELOG.rst
@@ -89,3 +89,4 @@ Changelog
 ====
 * Add multi-role example
 * Add inline role parameters
+* Fix blacklist roles bug